+++ test/Spec.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE DeriveGeneric #-}
 
 import BDSCOD.Conditioning
 import BDSCOD.Aggregation
@@ -6,7 +7,9 @@ import BDSCOD.Llhd
 import BDSCOD.Types
 import BDSCOD.Utility
 import Control.Monad (replicateM)
+import Data.List (sort)
 import Data.Maybe (fromJust, isJust)
+import Data.Tuple (swap)
 import qualified Data.Vector.Unboxed as Unboxed
 import qualified Epidemic as EpiSim
 import qualified Epidemic.BirthDeathSampling as EpiBDS
@@ -15,12 +18,15 @@ import Epidemic.Types.Events
 import Epidemic.Types.Parameter
 import Epidemic.Types.Population
 import qualified Epidemic.Utility as EpiUtil
+import GHC.Generics
+import Generic.Random (genericArbitraryU)
 import Numeric.GSL.SimulatedAnnealing
 import Numeric.LinearAlgebra.HMatrix
-import System.Random.MWC
+import qualified System.Random.MWC as MWC
 import Test.Hspec
 import Test.Hspec.QuickCheck
 import Test.QuickCheck
+import Test.QuickCheck.Gen
 import Test.Hspec.Core.QuickCheck (modifyMaxDiscardRatio)
 
 -- | Check if @y@ is withing @delta@ of @x@
@@ -30,6 +36,17 @@ withinDeltaOf :: (Ord a, Num a)
               -> a -- ^ x
               -> Bool
 withinDeltaOf delta y x = abs (y - x) < delta
+
+-- | Apply the @withinDeltaOf@ function to two lists.
+allWithinDeltaOf :: (Ord a, Num a) => a -> [a] -> [a] -> Bool
+allWithinDeltaOf _ [] [] = True
+allWithinDeltaOf delta [y] [x] = withinDeltaOf delta y x
+allWithinDeltaOf delta (y:ys) (x:xs) = withinDeltaOf delta y x && allWithinDeltaOf delta ys xs
+allWithinDeltaOf _ _ _ = False
+
+
+
+
 
 -- | Approximate the derivative of @f@ at @x@ with a step of size @h@.
 finiteDifference :: Fractional a
@@ -270,15 +287,6 @@ testLogPdeStatistics = do
                            snd' (_,a,_) = a
                            thd' (_,_,a) = a
 
-foo lam delay nbMean nbVar = (fooUnlogged lam delay nbMean nbVar, (a,b,c))
-                     where params lam = (Parameters (lam / 10, 0.3, 0.3, Timed [(1000,0.5)], 0.6, Timed []))
-                           pdeSol nbStats = (PDESol (nbFromMAndV nbStats) 1)
-                           scaledDelay d = d / 50
-                           fooUnlogged lam delay nbMean nbVar = pdeStatistics (params lam) (scaledDelay delay) (pdeSol (nbMean,nbVar))
-                           fooLogged lam delay nbMean nbVar  = (logPdeStatistics (params lam) (scaledDelay delay) (pdeSol (nbMean,nbVar)))
-                           (lnA,lnB,lnC) = fooLogged lam delay nbMean nbVar
-                           (a,b,c) = (exp lnA, exp lnB, exp lnC)
-
 
 
 testp0 = do
@@ -507,49 +515,6 @@ testConversion = do
 
 
 
-testAggregationSequenced =
-  let p1 = Person 1
-      p2 = Person 2
-      p3 = Person 3
-      p4 = Person 4
-      allEpiEvents = -- the events of the full epidemic
-        [ Infection 1 p1 p2
-        , Sampling 2 p1
-        , Infection 3 p2 p3
-        , Infection 4 p2 p4
-        , Removal 5 p4
-        , Sampling 6 p3
-        , Sampling 7 p2
-        ]
-      filteredEpiEvents = fromJust $ EpiBDSCOD.observedEvents allEpiEvents -- filter for just the observed epidemic events.
-      obs = -- the observations without aggregation
-        [ (1.0, OBirth)
-        , (1.0, ObsUnscheduledSequenced)
-        , (1.0, OBirth)
-        , (3.0, ObsUnscheduledSequenced)
-        , (1.0, ObsUnscheduledSequenced)
-        ]
-      obs' = eventsAsObservations filteredEpiEvents
-      aggTimes = fromJust $ maybeAggregationTimes [3.5,7.5]
-      aggObsObs =
-        [ (1.0, OBirth)
-        , (2.0, OBirth)
-        , (0.5, OCatastrophe 1)
-        , (4.0, OCatastrophe 2)
-        ]
-      aggObs = AggregatedObservations aggTimes aggObsObs
-      maybeAggObs' = aggregateUnscheduledObservations aggTimes obs
-    in
-    describe "Test aggregation definitions (only sequenced observations)" $ do
-      it "Observations are generated correctly" $ do
-          obs == obs' `shouldBe` True
-          allEpiEvents /= filteredEpiEvents `shouldBe` True
-      it "Aggregation works correctly" $ do
-        isJust maybeAggObs' `shouldBe` True
-        aggObs == fromJust maybeAggObs' `shouldBe` True
-
-
-
 testImpossibleParameters = do
   describe "Test correct handling of impossible parameters" $ do
     it "Test negative birth rate is impossible" $
@@ -606,7 +571,7 @@ bdsSimulations simRates simDuration =
           (\x -> probabilityUnobserved simRates (simDuration + x))
           [-0.1, 0.0, 0.1]
       probUnobserved = probUnobservedVals !! 1
-      numReplicates = 100
+      numReplicates = 1000
       numObservations = length . filter tmpIsSampling
       phat ns =
         (fromIntegral . length $ filter (> 0) ns) / (fromIntegral (length ns))
@@ -668,33 +633,147 @@ testMWCSeeding :: SpecWith ()
 testMWCSeeding = do
   describe "Testing MWC seeding" $ do
     it "test create works as expected" $ do
-      gen <- create
-      x1 <- (uniform gen :: IO Double)
-      x2 <- (uniform gen :: IO Double)
-      gen' <- create
-      x1' <- (uniform gen' :: IO Double)
+      gen <- MWC.create
+      x1 <- (MWC.uniform gen :: IO Double)
+      x2 <- (MWC.uniform gen :: IO Double)
+      gen' <- MWC.create
+      x1' <- (MWC.uniform gen' :: IO Double)
       (x1 /= x2) `shouldBe` True
       (x1 == x1') `shouldBe` True
     it "test initialise works as expected" $ do
-      xGen <- create
-      x1 <- (uniform xGen :: IO Double)
-      yGen <- initialize (Unboxed.fromList [1,2,3])
-      y1 <- (uniform yGen :: IO Double)
+      xGen <- MWC.create
+      x1 <- (MWC.uniform xGen :: IO Double)
+      yGen <- MWC.initialize (Unboxed.fromList [1,2,3])
+      y1 <- (MWC.uniform yGen :: IO Double)
       (x1 /= y1) `shouldBe` True
-      y2 <- (uniform yGen :: IO Double)
+      y2 <- (MWC.uniform yGen :: IO Double)
       (y1 /= y2) `shouldBe` True
-      zGen <- initialize (Unboxed.fromList [1,2,3])
-      z1 <- (uniform zGen :: IO Double)
+      zGen <- MWC.initialize (Unboxed.fromList [1,2,3])
+      z1 <- (MWC.uniform zGen :: IO Double)
       (y1 == z1) `shouldBe` True
-      wGen <- initialize (Unboxed.fromList [3,2,1])
-      w1 <- (uniform wGen :: IO Double)
+      wGen <- MWC.initialize (Unboxed.fromList [3,2,1])
+      w1 <- (MWC.uniform wGen :: IO Double)
       (z1 /= w1) `shouldBe` True
+
+
+
+-- | Generate a random @ObservedEvent@
+qcRandomObservedEvent :: Gen ObservedEvent
+qcRandomObservedEvent = do
+  isUnscheduled <- chooseAny
+  if isUnscheduled
+    then elements [OBirth,ObsUnscheduledSequenced,OOccurrence]
+    else do isSequenced <- chooseAny
+            numLineages <- suchThat chooseAny (>0)
+            if isSequenced
+              then return (OCatastrophe numLineages)
+              else return (ODisaster numLineages)
+
+
+-- | Generate a random list of @observation@ values
+qcRandomObservations :: Gen [Observation]
+qcRandomObservations = do
+  duration <- suchThat chooseAny (>0) :: Gen Time
+  eventAbsTimes <- listOf1 (choose (0,duration))
+  let eats = sort (0:eventAbsTimes)
+      timeDeltas = [b - a | (a, b) <- zip (init eats) (tail eats)]
+      numEvents = length eventAbsTimes
+  eventTypes <- vectorOf numEvents qcRandomObservedEvent
+  return $ zip timeDeltas eventTypes
+
+
+-- | A copy of the @withinDeltaOf@ function specialised to
+-- @AggregatedObservations@ to test for approximate equality of aggregated
+-- observations.
+withinDeltaOfAggObs :: Double
+                    -> AggregatedObservations
+                    -> AggregatedObservations
+                    -> Bool
+withinDeltaOfAggObs delta (AggregatedObservations aggTimes obs) (AggregatedObservations aggTimes' obs') =
+  withinDeltaOfAggTimes delta aggTimes aggTimes' &&  allWithinDeltaOfObs delta obs obs'
+
+withinDeltaOfAggTimes :: Double -> AggregationTimes -> AggregationTimes -> Bool
+withinDeltaOfAggTimes delta (AggTimes ts) (AggTimes ts') =
+  let times = map fst ts
+      obsEvents = map snd ts
+      times' = map fst ts'
+      obsEvents' = map snd ts'
+      timesWithinDelta = allWithinDeltaOf delta times times'
+      observedEventsEqual = all (uncurry (withinDeltaOfObsEvent delta)) (zip obsEvents obsEvents')
+  in timesWithinDelta && observedEventsEqual
+
+withinDeltaOfObsEvent :: Double -> ObservedEvent -> ObservedEvent -> Bool
+withinDeltaOfObsEvent delta (ODisaster nl) (ODisaster nl') = withinDeltaOf delta nl nl'
+withinDeltaOfObsEvent delta (OCatastrophe nl) (OCatastrophe nl') = withinDeltaOf delta nl nl'
+withinDeltaOfObsEvent _ OBirth OBirth = True
+withinDeltaOfObsEvent _ ObsUnscheduledSequenced ObsUnscheduledSequenced = True
+withinDeltaOfObsEvent _ OOccurrence OOccurrence = True
+withinDeltaOfObsEvent _ _ _ = False
+
+withinDeltaOfObs :: Double -> Observation -> Observation -> Bool
+withinDeltaOfObs delta (t,oe) (t',oe') = withinDeltaOf delta t t' && withinDeltaOfObsEvent delta oe oe'
+
+allWithinDeltaOfObs :: Double -> [Observation] -> [Observation] -> Bool
+allWithinDeltaOfObs _ [] [] = True
+allWithinDeltaOfObs delta [y] [x] = withinDeltaOfObs delta y x
+allWithinDeltaOfObs delta (y:ys) (x:xs) = withinDeltaOfObs delta y x && allWithinDeltaOfObs delta ys xs
+allWithinDeltaOfObs _ _ _ = False
+
+
+
+testAggregation :: SpecWith ()
+testAggregation = do
+  describe "Testing Aggregation" $ do
+    let smallDelta = 1e-4
+        tinyDelta = 1e-6
+        emptyAggTimes = fromJust (maybeAggregationTimes [] [])
+        propertyIdentity obs = let aggObs1 = aggregateUnscheduledObservations emptyAggTimes obs
+                                   aggObs2 = AggregatedObservations emptyAggTimes obs
+                                  in withinDeltaOfAggObs smallDelta aggObs1 aggObs2
+        duration obs = sum $ map fst obs
+        propertyRemoveSeq obs = let dur = duration obs
+                                    ats = fromJust $ maybeAggregationTimes [dur + tinyDelta] []
+                                    (AggregatedObservations _ obs') = aggregateUnscheduledObservations ats obs
+                                in  not $ any isUnscheduledSequenced obs'
+        propertyRemoveUnseq obs = let dur = duration obs
+                                      ats = fromJust $ maybeAggregationTimes [] [dur + tinyDelta]
+                                      (AggregatedObservations _ obs') = aggregateUnscheduledObservations ats obs
+                                  in  not $ any isOccurrence obs'
+        propertyRemoveUnsched obs = let dur = duration obs
+                                        ats = fromJust $ maybeAggregationTimes [dur + tinyDelta] [dur + tinyDelta + 1.0]
+                                        (AggregatedObservations _ obs') = aggregateUnscheduledObservations ats obs
+                                    in  not (any isOccurrence obs') && not (any isUnscheduledSequenced obs')
+        propertyBirthsRemain obs = let dur = duration obs
+                                       numBs = length $ filter isBirth obs
+                                       ats = fromJust $ maybeAggregationTimes [dur + tinyDelta] [dur + tinyDelta + 1.0]
+                                       (AggregatedObservations _ obs') = aggregateUnscheduledObservations ats obs
+                                       numBs' = length $ filter isBirth obs'
+                                   in numBs == numBs'
+        propertyLineagesConst os = let dur = duration os
+                                       numSeq = sum $ map numSequenced os
+                                       numUnseq = sum $ map numUnsequenced os
+                                       ats = fromJust $ maybeAggregationTimes [dur + tinyDelta] [dur + tinyDelta + 1.0]
+                                       (AggregatedObservations _ os') = aggregateUnscheduledObservations ats os
+                                       numSeq' = sum $ map numSequenced os'
+                                       numUnseq' = sum $ map numUnsequenced os'
+                                   in withinDeltaOf smallDelta numSeq numSeq' && withinDeltaOf smallDelta numUnseq numUnseq'
+    it "without aggregation nothing changes" $ forAll qcRandomObservations propertyIdentity
+    it "sequenced aggregation removes all such unscheduled observations" $ forAll qcRandomObservations propertyRemoveSeq
+    it "unsequenced aggregation removes all such unscheduled observations" $ forAll qcRandomObservations propertyRemoveUnseq
+    it "aggregating both removes all relevent observations" $ forAll qcRandomObservations propertyRemoveUnsched
+    it "aggregating leaves birth observations unchanged" $ forAll qcRandomObservations propertyBirthsRemain
+    it "aggregating leaves the number of observed lineages unchanged" $ forAll qcRandomObservations propertyLineagesConst
+
 
 
 main :: IO ()
 main = hspec $ do
-  testTestingHelpers
+  -- ** slow tests **
   testNbPGF
+  testHmatrixUsage
+  testConditioningProbability
+  -- ** fast tests **
+  testTestingHelpers
   testPdeStatistics
   testp0
   testRr
@@ -703,11 +782,8 @@ main = hspec $ do
   testConversion
   testImpossibleParameters
   testInhomBDSLlhd
-  testConditioningProbability
-  testHmatrixUsage
   testParameterUpdate
   testMWCSeeding
-  testAggregationSequenced
   testLogPdeGF1
   testLogPdeGF2
   testLogPdeGFDash1
@@ -716,4 +792,4 @@ main = hspec $ do
   testLogPdeGFDashDash2
   testLogSumExp
   testLogPdeStatistics
-
+  testAggregation